+++ sqlprojection/option.go
@@ -2,5 +2,6 @@ package sqlprojection
 
 // An Option configures the optional behavior of an SQL projection.
 type Option struct {
+	applyToAdaptor      func(*adaptor)
 	applyToCandidateSet func(*candidateSet)
 }